requirements.txt
cityscapesscripts==2.2.0
cycler==0.10.0
gdown==4.2.0
humanfriendly==9.2
kiwisolver==1.2.0
kornia==0.5.8
matplotlib==3.4.2
numpy==1.19.2
opencv-python==4.4.0.46
pandas==1.1.3
Pillow==8.3.1
prettytable==2.1.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.4.1
scipy==1.6.3
seaborn==0.11.1
timm==0.3.2
tqdm==4.48.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
yapf==0.31.0
